Breaking change: Greatly improve types for rules and rulebooks by using generics
This is a breaking change.
The values send to the rule definition is now always a single variable where it was a rest param. this means that every rule with more than 1 params has to be updated.

This helps us to make rules that we know will work with the rulebook.