Removed cosmiconfig
A ruleboundrc file does not make a lot of sense. If a user wants to add a rule config file, that's fine. However, it should probably not be named `.ruleboundrc` but something more specfic to that implementation.

So I've removed Cosmiconfig, and with it, the out of the box ability to store rule config in a file or package.json. This functionality can be fully replaced by passing the config to the Rulebook constructor.